+++ nablaDFT/dataset/__init__.py
@@ -1,3 +1,2 @@
-from .nablaDFT_dataset import *
-from .hamiltonian_database import HamiltonianDatabase
+from .nablaDFT_dataset import NablaDFT
 from .hamiltonian_dataset import HamiltonianDataset
